+++ src/ts/ir/highlightToolbar.ts
@@ -14,8 +14,8 @@ export const highlightToolbar = (vditor: IVditor) => {
             return;
         }
 
-        removeCurrentToolbar(vditor.toolbar.elements, Constants.TOOLBARS);
-        enableToolbar(vditor.toolbar.elements, Constants.TOOLBARS);
+        removeCurrentToolbar(vditor.toolbar.elements, Constants.EDIT_TOOLBARS);
+        enableToolbar(vditor.toolbar.elements, Constants.EDIT_TOOLBARS);
 
         const range = getEditorRange(vditor.ir.element);
         let typeElement = range.startContainer as HTMLElement;