+++ src/ts/toolbar/Both.ts
@@ -1,3 +1,4 @@
+import {Constants} from "../constants";
 import {setPreviewMode} from "../ui/setPreviewMode";
 import {getEventName} from "../util/compatibility";
 import {MenuItem} from "./MenuItem";
@@ -9,11 +10,15 @@ export class Both extends MenuItem {
             this.element.children[0].classList.add("vditor-menu--current");
         }
         this.element.children[0].addEventListener(getEventName(), (event) => {
+            const btnElement = this.element.firstElementChild;
+            if (btnElement.classList.contains(Constants.CLASS_MENU_DISABLED)) {
+                return;
+            }
             event.preventDefault();
             if (vditor.currentMode !== "sv") {
                 return;
             }
-            if (vditor.currentPreviewMode === "both") {
+            if (vditor.options.preview.mode === "both") {
                 setPreviewMode("editor", vditor);
             } else {
                 setPreviewMode("both", vditor);