+++ src/ts/ui/setPreviewMode.ts
@@ -1,11 +1,11 @@
 import {removeCurrentToolbar} from "../toolbar/setToolbar";
 import {setCurrentToolbar} from "../toolbar/setToolbar";
 
-export const setPreviewMode = (mode: keyof IPreviewMode, vditor: IVditor) => {
-    if (vditor.currentPreviewMode === mode) {
+export const setPreviewMode = (mode: "both" | "editor", vditor: IVditor) => {
+    if (vditor.options.preview.mode === mode) {
         return;
     }
-    vditor.currentPreviewMode = mode;
+    vditor.options.preview.mode = mode;
 
     switch (mode) {
         case "both":
@@ -14,26 +14,12 @@ export const setPreviewMode = (mode: keyof IPreviewMode, vditor: IVditor) => {
             vditor.preview.render(vditor);
 
             setCurrentToolbar(vditor.toolbar.elements, ["both"]);
-            removeCurrentToolbar(vditor.toolbar.elements, ["preview"]);
-
             break;
         case "editor":
             vditor.sv.element.style.display = "block";
             vditor.preview.element.style.display = "none";
 
-            removeCurrentToolbar(vditor.toolbar.elements, ["preview"]);
             removeCurrentToolbar(vditor.toolbar.elements, ["both"]);
-
-            break;
-        case "preview":
-            vditor.sv.element.style.display = "none";
-            vditor.preview.element.style.display = "block";
-            vditor.preview.render(vditor);
-            vditor.sv.element.blur();
-
-            setCurrentToolbar(vditor.toolbar.elements, ["preview"]);
-            removeCurrentToolbar(vditor.toolbar.elements, ["both"]);
-
             break;
         default:
             break;